+++ src-tauri/src/export/cbz.rs
@@ -1,0 +1,296 @@
+use std::{
+    io::Write,
+    sync::{atomic::AtomicU32, Arc},
+};
+
+use anyhow::{anyhow, Context};
+use rayon::iter::{IntoParallelIterator, ParallelIterator};
+use tauri::AppHandle;
+use tauri_specta::Event;
+use zip::{write::SimpleFileOptions, ZipWriter};
+
+use crate::{
+    config::ExportSkipMode,
+    events::ExportCbzEvent,
+    export::{
+        get_downloaded_chapters, get_downloaded_chapters_by_uuids, get_image_paths,
+        ComicExportLockGuard, ExportFormat,
+    },
+    extensions::AppHandleExt,
+    types::{ChapterInfo, Comic, ComicInfo},
+};
+
+struct CbzErrorEventGuard {
+    uuid: String,
+    app: AppHandle,
+    success: bool,
+}
+
+impl Drop for CbzErrorEventGuard {
+    fn drop(&mut self) {
+        if self.success {
+            return;
+        }
+
+        let uuid = self.uuid.clone();
+        let _ = ExportCbzEvent::Error { uuid }.emit(&self.app);
+    }
+}
+
+// ============================================================================
+// CBZ 导出函数
+// ============================================================================
+
+/// 公开接口：导出全部已下载章节为CBZ
+#[allow(clippy::cast_possible_wrap)]
+#[allow(clippy::cast_possible_truncation)]
+#[allow(clippy::too_many_lines)]
+pub fn cbz(app: &AppHandle, comic: &Comic) -> anyhow::Result<()> {
+    let comic_path_word = &comic.comic.path_word;
+    let comic_title = &comic.comic.name;
+    let export_lock = app.get_export_lock().inner().clone();
+
+    // 检查导出锁
+    if !export_lock.try_acquire(comic_path_word) {
+        return Err(anyhow!("漫画`{comic_title}`正在导出，请稍后再试"));
+    }
+
+    let _guard = ComicExportLockGuard {
+        lock: export_lock.clone(),
+        path_word: comic_path_word.clone(),
+    };
+
+    // 获取配置
+    let skip_mode = app.get_config().read().export_skip_mode;
+
+    // 获取已下载章节
+    let downloaded_chapters = get_downloaded_chapters(&comic.comic.groups);
+
+    // 调用内部实现
+    cbz_internal(
+        app,
+        comic,
+        downloaded_chapters,
+        skip_mode,
+        comic_path_word,
+        comic_title,
+    )
+}
+
+/// 公开接口：导出指定已下载章节为CBZ
+#[allow(clippy::cast_possible_wrap)]
+#[allow(clippy::cast_possible_truncation)]
+#[allow(clippy::too_many_lines)]
+pub fn cbz_chapters(
+    app: &AppHandle,
+    comic: &Comic,
+    chapter_uuids: Vec<String>,
+) -> anyhow::Result<()> {
+    let comic_path_word = &comic.comic.path_word;
+    let comic_title = &comic.comic.name;
+    let export_lock = app.get_export_lock().inner().clone();
+
+    // 检查导出锁
+    if !export_lock.try_acquire(comic_path_word) {
+        return Err(anyhow!("漫画`{comic_title}`正在导出，请稍后再试"));
+    }
+
+    let _guard = ComicExportLockGuard {
+        lock: export_lock.clone(),
+        path_word: comic_path_word.clone(),
+    };
+
+    // 获取指定章节（用户主动选择，不跳过）
+    let downloaded_chapters = get_downloaded_chapters_by_uuids(&comic.comic.groups, &chapter_uuids);
+
+    // 调用内部实现，skip_mode = None（用户主动选择，不跳过）
+    cbz_internal(
+        app,
+        comic,
+        downloaded_chapters,
+        ExportSkipMode::None,
+        comic_path_word,
+        comic_title,
+    )
+}
+
+/// 内部实现：导出CBZ
+#[allow(clippy::cast_possible_wrap)]
+#[allow(clippy::cast_possible_truncation)]
+#[allow(clippy::too_many_lines)]
+fn cbz_internal(
+    app: &AppHandle,
+    comic: &Comic,
+    downloaded_chapters: Vec<ChapterInfo>,
+    skip_mode: ExportSkipMode,
+    comic_path_word: &str,
+    comic_title: &str,
+) -> anyhow::Result<()> {
+    // 用于生成格式化的xml
+    let xml_cfg = yaserde::ser::Config {
+        perform_indent: true,
+        ..Default::default()
+    };
+    let event_uuid = uuid::Uuid::new_v4().to_string();
+
+    // 发送开始导出cbz事件
+    let _ = ExportCbzEvent::Start {
+        uuid: event_uuid.clone(),
+        comic_title: comic_title.to_string(),
+        total: downloaded_chapters.len() as u32,
+    }
+    .emit(app);
+
+    // 如果success为false，drop时发送Error事件
+    let mut error_event_guard = CbzErrorEventGuard {
+        uuid: event_uuid.clone(),
+        app: app.clone(),
+        success: false,
+    };
+
+    // 用来记录导出进度
+    let current = Arc::new(AtomicU32::new(0));
+
+    let extension = ExportFormat::Cbz.extension();
+    let comic_export_dir = comic
+        .get_comic_export_dir(app)
+        .context(format!("`{comic_title}` 获取导出目录失败"))?;
+    let cbz_export_dir = comic_export_dir.join(extension);
+
+    // 并发处理
+    let download_chapters = downloaded_chapters.into_par_iter();
+    download_chapters.try_for_each(|mut chapter_info| -> anyhow::Result<()> {
+        let chapter_title = &chapter_info.chapter_title;
+        let group_name = &chapter_info.group_name;
+        let err_prefix = format!("`{comic_title} - {group_name} - {chapter_title}`");
+
+        // 获取导出路径
+        let chapter_download_dir = chapter_info
+            .chapter_download_dir
+            .as_ref()
+            .context(format!("{err_prefix} `chapter_download_dir`字段为`None`"))?;
+        let chapter_download_dir_name = chapter_download_dir
+            .file_name()
+            .and_then(|name: &std::ffi::OsStr| name.to_str())
+            .context(format!(
+                "{err_prefix} 获取`{}`的目录名失败",
+                chapter_download_dir.display()
+            ))?;
+        let chapter_relative_dir = chapter_info
+            .get_chapter_relative_dir(comic)
+            .context(format!("{err_prefix} 获取章节相对目录失败"))?;
+        let chapter_relative_dir_parent = chapter_relative_dir.parent().context(format!(
+            "{err_prefix} `{}`没有父目录",
+            chapter_relative_dir.display()
+        ))?;
+        let chapter_export_dir = cbz_export_dir.join(chapter_relative_dir_parent);
+
+        // 保证导出目录存在
+        std::fs::create_dir_all(&chapter_export_dir).context(format!(
+            "{err_prefix} 创建目录`{}`失败",
+            chapter_export_dir.display()
+        ))?;
+
+        let zip_path = chapter_export_dir.join(format!("{chapter_download_dir_name}.{extension}"));
+
+        // 跳过逻辑
+        let should_skip = match skip_mode {
+            ExportSkipMode::SkipExported if chapter_info.is_cbz_exported => true,
+            ExportSkipMode::SkipExisting if zip_path.exists() => true,
+            _ => false,
+        };
+
+        if should_skip {
+            // 更新进度
+            let current = current.fetch_add(1, std::sync::atomic::Ordering::Relaxed) + 1;
+            let _ = ExportCbzEvent::Progress {
+                uuid: event_uuid.clone(),
+                current,
+            }
+            .emit(app);
+            return Ok(());
+        }
+
+        // 生成ComicInfo
+        let comic_info = ComicInfo::from(comic, &chapter_info);
+        // 序列化ComicInfo为xml
+        let comic_info_xml = yaserde::ser::to_string_with_config(&comic_info, &xml_cfg)
+            .map_err(|err_msg| anyhow!("{err_prefix} 序列化`ComicInfo.xml`失败: {err_msg}"))?;
+
+        // 创建cbz文件
+        let zip_file = std::fs::File::create(&zip_path)
+            .context(format!("{err_prefix} 创建文件`{}`失败", zip_path.display()))?;
+        let mut zip_writer = ZipWriter::new(zip_file);
+
+        // 把ComicInfo.xml写入cbz
+        zip_writer
+            .start_file("ComicInfo.xml", SimpleFileOptions::default())
+            .context(format!(
+                "{err_prefix} 在`{}`创建`ComicInfo.xml`失败",
+                zip_path.display()
+            ))?;
+        zip_writer
+            .write_all(comic_info_xml.as_bytes())
+            .context(format!("{err_prefix} 写入`ComicInfo.xml`失败"))?;
+
+        let image_paths = get_image_paths(chapter_download_dir).context(format!(
+            "{err_prefix} 获取`{}`中的图片失败",
+            chapter_download_dir.display()
+        ))?;
+
+        for image_path in image_paths {
+            let filename = image_path
+                .file_name()
+                .and_then(|name: &std::ffi::OsStr| name.to_str())
+                .context(format!(
+                    "{err_prefix} 获取`{}`的文件名失败",
+                    image_path.display()
+                ))?;
+            // 将文件写入cbz
+            zip_writer
+                .start_file(filename, SimpleFileOptions::default())
+                .context(format!(
+                    "{err_prefix} 在`{}`创建`{filename:?}`失败",
+                    zip_path.display()
+                ))?;
+            let mut file = std::fs::File::open(&image_path)
+                .context(format!("{err_prefix} 打开`{}`失败", image_path.display()))?;
+            std::io::copy(&mut file, &mut zip_writer).context(format!(
+                "{err_prefix} 将`{}`写入`{}`失败",
+                image_path.display(),
+                zip_path.display()
+            ))?;
+        }
+
+        zip_writer
+            .finish()
+            .context(format!("{err_prefix} 关闭`{}`失败", zip_path.display()))?;
+
+        // 更新章节导出状态
+        chapter_info.is_cbz_exported = true;
+        chapter_info.save_metadata()?;
+
+        // 更新导出cbz的进度
+        let current = current.fetch_add(1, std::sync::atomic::Ordering::Relaxed) + 1;
+        // 发送导出cbz进度事件
+        let _ = ExportCbzEvent::Progress {
+            uuid: event_uuid.clone(),
+            current,
+        }
+        .emit(app);
+
+        Ok(())
+    })?;
+
+    // 标记为成功，后面drop时就不会发送Error事件
+    error_event_guard.success = true;
+    // 发送导出cbz完成事件
+    let _ = ExportCbzEvent::End {
+        uuid: event_uuid,
+        comic_path_word: comic_path_word.to_string(),
+        chapter_export_dir: cbz_export_dir,
+    }
+    .emit(app);
+
+    Ok(())
+}
